Move ls export to TemplateManager
