+++ src/main/java/com/redhat/devtools/lsp4ij/launching/templates/LanguageServerTemplateManager.java
@@ -11,19 +11,30 @@
 package com.redhat.devtools.lsp4ij.launching.templates;
 
 import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
 import com.intellij.openapi.application.ApplicationManager;
+import com.intellij.openapi.vfs.VirtualFile;
 import com.redhat.devtools.lsp4ij.internal.IntelliJPlatformUtils;
+import com.redhat.devtools.lsp4ij.server.definition.LanguageServerDefinition;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import com.redhat.devtools.lsp4ij.internal.StringUtils;
+import com.redhat.devtools.lsp4ij.server.definition.launching.UserDefinedLanguageServerDefinition;
 
-import java.io.BufferedInputStream;
+import java.io.ByteArrayOutputStream;
+import java.util.zip.ZipEntry;
+import java.util.zip.ZipOutputStream;
 import java.io.IOException;
+import java.nio.charset.StandardCharsets;
+import java.io.BufferedInputStream;
 import java.io.InputStreamReader;
 import java.io.Reader;
 import java.util.Collections;
 import java.util.List;
+
+import static com.redhat.devtools.lsp4ij.launching.templates.LanguageServerTemplate.*;
 
 /**
  * Language server template manager.
@@ -35,6 +46,8 @@ public class LanguageServerTemplateManager {
     private static final String TEMPLATES_DIR = "templates";
 
     private final List<LanguageServerTemplate> templates;
+
+    private String lsName;
 
     public static LanguageServerTemplateManager getInstance() {
         return ApplicationManager.getApplication().getService(LanguageServerTemplateManager.class);
@@ -61,5 +74,64 @@ public class LanguageServerTemplateManager {
     static Reader loadTemplateReader(@NotNull String path) {
         var is = LanguageServerTemplateManager.class.getClassLoader().getResourceAsStream(TEMPLATES_DIR + "/" + path);
         return is !=null? new InputStreamReader(new BufferedInputStream(is)) : null;
+    }
+
+    public void importLsTemplate(VirtualFile templateZip) {
+
+    }
+
+    public void exportLsTemplates(@NotNull VirtualFile exportZip, @NotNull List<LanguageServerDefinition> lsDefinitions) {
+        ApplicationManager.getApplication().runWriteAction(() -> {
+            try {
+                exportZip.setBinaryContent(createZipFromLanguageServers(lsDefinitions));
+            } catch (IOException ex) {
+                LOGGER.warn(ex.getLocalizedMessage(), ex);
+            }
+        });
+    }
+
+    /**
+     * Creates a zip file by handling each user defined language server definitions
+     * @return zip file as a byte array
+     */
+    private byte[] createZipFromLanguageServers(@NotNull List<LanguageServerDefinition> lsDefinitions) throws IOException {
+        ByteArrayOutputStream baos = new ByteArrayOutputStream();
+        ZipOutputStream zos = new ZipOutputStream(baos);
+
+        for (LanguageServerDefinition lsDefinition : lsDefinitions) {
+            Gson gson = new GsonBuilder()
+                    .registerTypeAdapter(UserDefinedLanguageServerDefinition.class, new LanguageServerDefinitionSerializer())
+                    .setPrettyPrinting()
+                    .create();
+            String template = gson.toJson(lsDefinition);
+            String initializationOptions = ((UserDefinedLanguageServerDefinition) lsDefinition).getInitializationOptionsContent();
+            String settings = ((UserDefinedLanguageServerDefinition) lsDefinition).getConfigurationContent();
+            lsName = lsDefinition.getDisplayName();
+
+            writeToZip(TEMPLATE, template, zos);
+            writeToZip(INITIALIZATION_OPTIONS, initializationOptions, zos);
+            writeToZip(SETTINGS, settings, zos);
+            zos.closeEntry();
+        }
+
+        zos.close();
+        return baos.toByteArray();
+    }
+
+    /**
+     * Writes a file (name + content) to a zip output stream
+     * @param filename name of the file to write
+     * @param content file content
+     * @param zos to write the file to
+     */
+    private void writeToZip(String filename, String content, ZipOutputStream zos) throws IOException {
+        if (StringUtils.isBlank(content)) {
+            content = "{}";
+        }
+
+        ZipEntry entry = new ZipEntry(lsName + "/" + filename);
+        zos.putNextEntry(entry);
+        zos.write(content.getBytes(StandardCharsets.UTF_8));
+        zos.closeEntry();
     }
 }
