+++ src/main/resources/messages/LanguageServerBundle.properties
@@ -88,9 +88,9 @@ action.lsp.console.explorer.delete.server.confirm.dialog.message=Do you want to 
 action.lsp.console.explorer.edit.server.text=Edit Server
 action.lsp.console.explorer.edit.server.description=Edit the selected language server
 action.lsp.console.explorer.export.server.text=Export Server
-action.lsp.console.explorer.export.server.description=Export the selected language server as json
+action.lsp.console.explorer.export.server.description=Export the selected language server
 action.lsp.console.explorer.export.servers.text=Export Servers
-action.lsp.console.explorer.export.servers.description=Export the selected language servers as json
+action.lsp.console.explorer.export.servers.description=Export the selected language servers
 action.lsp.console.explorer.export.servers.zip.save.title=Save LS Definition
 action.lsp.console.explorer.export.servers.zip.save.description=Choose location for the LS definition export
 action.lsp.console.folding.text=Collapse/Expand All
